pyproject.toml
[project]
name = "tech_read_bot"
version = "0.1.0"
description = "Discord bot for managing reading groups and discussions"
authors = [
    {name = "pongpatapee", email = "pongpatapee1021@gmail.com"},
]
dependencies = [
    "discord-py>=2.5.2",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "pydantic>=2.11.7",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
